moving proxy to its own middleware removing cors as its not needed
