+++ package.json
@@ -20,7 +20,6 @@
     "@quasar/extras": "^1.16.12",
     "@vueuse/core": "^10.7.0",
     "axios": "^1.2.1",
-    "cors": "^2.8.5",
     "d3-interpolate": "^3.0.1",
     "dayjs": "^1.11.11",
     "dompurify": "^3.1.2",
@@ -48,7 +47,6 @@
     "@quasar/app-vite": "^1.10.2",
     "@quasar/vite-plugin": "^1.8.0",
     "@types/axios": "^0.14.0",
-    "@types/cors": "^2.8.17",
     "@types/dompurify": "^3.0.5",
     "@types/lodash-es": "^4.17.12",
     "@types/marked": "^5.0.1",
